fixed major bugs

+++ ex00/srcs/parsing/ft_parse.c
@@ -6,17 +6,17 @@
 /*   By: bsouleau <bsouleau@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/18 13:57:07 by bsouleau          #+#    #+#             */
-/*   Updated: 2021/09/18 13:57:16 by bsouleau         ###   ########lyon.fr   */
+/*   Updated: 2021/09/18 16:35:42 by bsouleau         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/ft_utils.h"
 
-int	ft_parse(char *str)
+long int	ft_parse(char *str)
 {
 	t_parsed		number;
-	int				first_parse;
-	unsigned int	output;
+	long int		first_parse;
+	long int		output;
 
 	first_parse = ft_atoi(str);
 	if (!(first_parse == -1))
